requirements.txt
pandas
numpy
requests
openai
schedule
ta
python-dotenv
python-binance
```

6. .env 파일 예시:
<antArtifact identifier="env-file" type="application/vnd.ant.code" language="text" title="Environment Variables for API Keys">
OPENAI_API_KEY=your_openai_api_key
GLASSNODE_API_KEY=your_glassnode_api_key
```

이러한 수정을 통해 코드의 안정성과 유연성을 높였습니다. 주요 변경 사항은:
1. Import 문 추가 및 수정
2. 메서드의 안전한 파라미터 접근
3. 예외 처리 강화
4. 기본 파라미터 추가
5. 환경 변수 관리를 위한 .env 파일 추가
6. 종속성 관리를 위한 requirements.txt 추가

실제 운용 시에는 추가적인 테스트와 리스크 관리가 필수적입니다.